+++ packages/yourloops/components/patient-list/patient-list-page.tsx
@@ -31,9 +31,7 @@ import { usePatientListHook } from './patient-list.hook'
 import { PatientListTabs } from './models/enums/patient-list.enum'
 import { GlobalStyles } from 'tss-react'
 import { useTheme } from '@mui/material/styles'
-import {
-  CurrentMedicalTeamPatientList
-} from './current-patient-list/current-medical-team-patient-list/current-medical-team-patient-list'
+import { MedicalTeamPatientList } from './current-patient-list/medical-team-patient-list/medical-team-patient-list'
 import { PendingPatientList } from './pending-patient-list/pending-patient-list'
 import { setPageTitle } from '../../lib/utils'
 import { useTranslation } from 'react-i18next'
@@ -41,8 +39,8 @@ import { useAuth } from '../../lib/auth'
 import { useSelectedTeamContext } from '../../lib/selected-team/selected-team.provider'
 import TeamUtils from '../../lib/team/team.util'
 import {
-  CurrentPrivateTeamOrCaregiverPatientList
-} from './current-patient-list/current-private-team-or-caregiver-patient-list/current-private-team-or-caregiver-patient-list'
+  PrivateTeamOrCaregiverPatientList
+} from './current-patient-list/private-team-or-caregiver-patient-list/private-team-or-caregiver-patient-list'
 
 export const PatientListPage: FunctionComponent = () => {
   const theme = useTheme()
@@ -73,10 +71,10 @@ export const PatientListPage: FunctionComponent = () => {
       />
 
       {selectedTab === PatientListTabs.Current && isCaregiverUserOrPrivateTeam &&
-        <CurrentPrivateTeamOrCaregiverPatientList patients={patients} />
+        <PrivateTeamOrCaregiverPatientList patients={patients} />
       }
       {selectedTab === PatientListTabs.Current && !isCaregiverUserOrPrivateTeam &&
-        <CurrentMedicalTeamPatientList patients={patients} />
+        <MedicalTeamPatientList patients={patients} />
       }
       {selectedTab === PatientListTabs.Pending &&
         <PendingPatientList patients={patients} />
